pull the nesting out to its own node
